+++ branch-alpha/test.mjs
@@ -35,9 +35,10 @@ import {
     childProcessSpawn2,
     ciBuildExt,
     dbCloseAsync,
-    dbExecAndReturnFirstRow,
-    dbExecAndReturnFirstTable,
-    dbExecAndReturnLastBlobAsync,
+    dbExecAndReturnLastBlob,
+    dbExecAndReturnLastRow,
+    dbExecAndReturnLastTable,
+    dbExecAndReturnLastValue,
     dbExecAsync,
     dbFileLoadAsync,
     dbFileSaveAsync,
@@ -200,18 +201,19 @@ jstestDescribe((
                 ];
                 assertJsonEqual(bufActual, bufExpect, {
                     bufActual,
+                    bufExpect,
                     ii,
                     valExpect,
                     valIn
                 });
             }));
         }
-        async function test_dbBind_lastblob(ii, valIn, valExpect) {
+        async function test_dbBind_lastBlob(ii, valIn, valExpect) {
             let bufActual;
             let bufExpect;
             if (valExpect === Error) {
                 assertErrorThrownAsync(
-                    dbExecAndReturnLastBlobAsync.bind(undefined, {
+                    dbExecAndReturnLastBlob.bind(undefined, {
                         bindList: [valIn],
                         db,
                         sql: "SELECT 1, 2, 3; SELECT 1, 2, ?"
@@ -221,7 +223,7 @@ jstestDescribe((
                 return;
             }
             bufActual = new TextDecoder().decode(
-                await dbExecAndReturnLastBlobAsync({
+                await dbExecAndReturnLastBlob({
                     bindList: [valIn],
                     db,
                     sql: "SELECT 1, 2, 3; SELECT 1, 2, ?"
@@ -279,13 +281,39 @@ jstestDescribe((
                 valIn
             });
         }
+        async function test_dbBind_lastValue(ii, valIn, valExpect) {
+            let valActual;
+            if (valExpect === Error) {
+                assertErrorThrownAsync(
+                    dbExecAndReturnLastValue.bind(undefined, {
+                        bindList: [valIn],
+                        db,
+                        sql: "SELECT 1, 2, 3; SELECT 1, 2, ?"
+                    }),
+                    "inclusive-range|not JSON serializable"
+                );
+                return;
+            }
+            valActual = await dbExecAndReturnLastValue({
+                bindList: [valIn],
+                db,
+                sql: "SELECT 1, 2, 3; SELECT 1, 2, ?"
+            });
+            assertJsonEqual(valActual, valExpect, {
+                ii,
+                valActual,
+                valExpect,
+                valIn
+            });
+        }
         async function test_dbBind_responseType(ii, valIn, valExpect) {
             await Promise.all([
                 "arraybuffer",
                 "list",
+                "lastvalue",
                 undefined
             ].map(async function (responseType) {
-                let bufActual;
+                let valActual;
                 if (valExpect === Error) {
                     assertErrorThrownAsync(
                         dbExecAsync.bind(undefined, {
@@ -298,7 +326,7 @@ jstestDescribe((
                     );
                     return;
                 }
-                bufActual = await dbExecAsync({
+                valActual = await dbExecAsync({
                     bindList: [valIn],
                     db,
                     responseType,
@@ -306,20 +334,22 @@ jstestDescribe((
                 });
                 switch (responseType) {
                 case "arraybuffer":
-                    bufActual = JSON.parse(
-                        new TextDecoder().decode(bufActual)
+                    valActual = JSON.parse(
+                        new TextDecoder().decode(valActual)
                     )[0][1][0];
                     break;
+                case "lastvalue":
+                    break;
                 case "list":
-                    bufActual = bufActual[0][1][0];
+                    valActual = valActual[0][1][0];
                     break;
                 default:
-                    bufActual = bufActual[0][0].val;
+                    valActual = valActual[0][0].val;
                 }
-                assertJsonEqual(bufActual, valExpect, {
-                    bufActual,
+                assertJsonEqual(valActual, valExpect, {
                     ii,
                     responseType,
+                    valActual,
                     valExpect,
                     valIn
                 });
@@ -379,9 +409,11 @@ jstestDescribe((
             [{}, "{}"],
             // 6. string
             ["", ""],
+            ["'", "'"],
             ["0", "0"],
             ["1", "1"],
             ["2", "2"],
+            ["\"", "\""],
             ["\u0000", "\u0000"],
             ["\u0000\u{1f600}\u0000", "\u0000\u{1f600}\u0000"],
             ["a".repeat(9999), "a".repeat(9999)],
@@ -392,7 +424,8 @@ jstestDescribe((
         ].map(async function ([valIn, valExpect], ii) {
             await Promise.all([
                 test_dbBind_exec(ii, valIn, valExpect),
-                test_dbBind_lastblob(ii, valIn, valExpect),
+                test_dbBind_lastBlob(ii, valIn, valExpect),
+                test_dbBind_lastValue(ii, valIn, valExpect),
                 test_dbBind_responseType(ii, valIn, valExpect)
             ]);
         }));
@@ -459,9 +492,11 @@ jstestDescribe((
             [{}, Error],
             // 6. string
             ["", ""],
+            ["'", "'"],
             ["0", "0"],
             ["1", "1"],
             ["2", "2"],
+            ["\"", "\""],
             ["\u0000", ""],
             ["\u0000\u{1f600}\u0000", "\u0000\u{1f600}"],
             ["a".repeat(9999), "a".repeat(9999)],
@@ -521,45 +556,65 @@ jstestDescribe((
         let db = await dbOpenAsync({
             filename: ":memory:"
         });
-        // test dbExecAndReturnFirstRow null-case handling-behavior
+        // test dbExecAndReturnLastRow null-case handling-behavior
         assertJsonEqual(
             noop(
-                await dbExecAndReturnFirstRow({
+                await dbExecAndReturnLastRow({
                     db,
                     sql: "SELECT 0 WHERE 0"
                 })
             ),
             {}
         );
-        // test dbExecAndReturnFirstTable null-case handling-behavior
+        // test dbExecAndReturnLastTable null-case handling-behavior
         assertJsonEqual(
             noop(
-                await dbExecAndReturnFirstTable({
+                await dbExecAndReturnLastTable({
                     db,
                     sql: "SELECT 0 WHERE 0"
                 })
             ),
             []
         );
-        // test dbExecAndReturnLastBlobAsync null-case handling-behavior
+        // test dbExecAndReturnLastBlob null-case handling-behavior
         assertJsonEqual(
             new TextDecoder().decode(
-                await dbExecAndReturnLastBlobAsync({
+                await dbExecAndReturnLastBlob({
                     db,
                     sql: "SELECT 0 WHERE 0"
                 })
             ),
             ""
         );
-        // test dbExecAndReturnLastBlobAsync string handling-behavior
+        // test dbExecAndReturnLastBlob string handling-behavior
         assertJsonEqual(
             new TextDecoder().decode(
-                await dbExecAndReturnLastBlobAsync({
+                await dbExecAndReturnLastBlob({
                     db,
-                    sql: "SELECT 1234"
+                    sql: "SELECT 1, 2, 3"
                 })
             ),
-            "1234"
+            "3"
+        );
+        // test dbExecAndReturnLastValue null-case handling-behavior
+        assertJsonEqual(
+            noop(
+                await dbExecAndReturnLastValue({
+                    db,
+                    sql: "SELECT 0 WHERE 0"
+                })
+            ),
+            null
+        );
+        // test dbExecAndReturnLastValue json handling-behavior
+        assertJsonEqual(
+            noop(
+                await dbExecAndReturnLastValue({
+                    db,
+                    sql: "SELECT 1, 2, 3"
+                })
+            ),
+            3
         );
     });
     jstestIt((
@@ -687,13 +742,7 @@ SELECT * FROM testDbExecAsync2;
             db,
             sql: "SELECT * FROM t01"
         });
-        assertJsonEqual(data, [
-            [
-                {
-                    c01: 1
-                }
-            ]
-        ]);
+        assertJsonEqual(data, [[{c01: 1}]]);
     });
     jstestIt((
         "test dbOpenAsync handling-behavior"
@@ -790,6 +839,51 @@ jstestDescribe((
 });
 
 jstestDescribe((
+    "test_lgbm"
+), function test_lgbm() {
+    jstestIt((
+        "test lgbm handling-behavior"
+    ), async function () {
+        let data;
+        let db = await dbOpenAsync({filename: ":memory:"});
+        dbExecAsync({
+            db,
+            sql: (`
+SELECT lgbm_dlopen(NULL);
+
+DROP TABLE IF EXISTS __tmp1;
+CREATE TEMP TABLE __tmp1 AS
+    SELECT
+        lgbm_datasetcreatefromfile(
+            'test_lgbm_binary.train',
+            'max_bin=15',
+            0
+        ) AS handle;
+
+SELECT
+        lgbm_datasetdumptext(handle, '.tmp/test_lgbm_datasetdump.txt')
+    FROM __tmp1;
+            `)
+        });
+        data = await dbExecAndReturnLastRow({
+            db,
+            sql: (`
+SELECT
+        handle,
+        lgbm_datasetgetnumdata(handle) AS num_data,
+        lgbm_datasetgetnumfeature(handle) AS num_feature
+    FROM __tmp1;
+            `)
+        });
+        debugInline(data);
+        await dbExecAsync({
+            db,
+            sql: `SELECT lgbm_datasetfree(${data.handle});`
+        });
+    });
+});
+
+jstestDescribe((
     "test_misc"
 ), function test_misc() {
     jstestIt((
@@ -835,7 +929,7 @@ jstestDescribe((
         assertJsonEqual(
             "not an error",
             noop(
-                await dbExecAndReturnFirstRow({
+                await dbExecAndReturnLastRow({
                     db,
                     sql: `SELECT throwerror(0) AS val`
                 })
@@ -917,7 +1011,7 @@ jstestDescribe((
             let valActual;
             try {
                 valActual = noop(
-                    await dbExecAndReturnFirstRow({
+                    await dbExecAndReturnLastRow({
                         bindList: {
                             valIn
                         },
@@ -1168,16 +1262,12 @@ SELECT
             ]) {
                 let sql = `SELECT ${func}(${arg}) AS val`;
                 let valActual = noop(
-                    await dbExecAndReturnFirstRow({
+                    await dbExecAndReturnLastRow({
                         db,
                         sql
                     })
                 ).val;
-                assertJsonEqual(valActual, valExpect, {
-                    sql,
-                    valActual,
-                    valExpect
-                });
+                assertJsonEqual(valActual, valExpect);
             });
         });
     });
@@ -1188,7 +1278,7 @@ SELECT
             filename: ":memory:"
         });
         await (async function () {
-            let valActual = await dbExecAndReturnFirstTable({
+            let valActual = await dbExecAndReturnLastTable({
                 db,
                 sql: (`
 -- test null-case handling-behavior
@@ -1302,19 +1392,20 @@ SELECT
                 Array.from(data).reverse()
             ].map(async function (data) {
                 let valActual;
-                valActual = await dbExecAndReturnFirstRow({
+                valActual = await dbExecAndReturnLastRow({
                     bindList: {
                         tmp1: JSON.stringify(data)
                     },
                     db,
                     sql: (`
+-- test null-case handling-behavior
+SELECT quantile(value, ${kk}) AS qnt FROM JSON_EACH($tmp1) WHERE 0;
+-- test last-row handling-behavior
 SELECT
         median(value) AS mdn,
         quantile(value, ${kk}) AS qnt,
         ROUND(stdev(value), 8) AS std
     FROM JSON_EACH($tmp1);
--- test null-case handling-behavior
-SELECT quantile(value, ${kk}) AS qnt FROM JSON_EACH($tmp1) WHERE 0;
                     `)
                 });
                 assertJsonEqual(
@@ -2254,7 +2345,7 @@ SELECT doublearray_jsonto(win_sinefit2(1, 2, 3, 4)) FROM __tmp1;
             // test win_sinefit2-aggregate-normal handling-behavior
             (async function () {
                 let valActual;
-                valActual = await dbExecAndReturnFirstRow({
+                valActual = await dbExecAndReturnLastRow({
                     bindList: {
                         valIn
                     },
@@ -2311,7 +2402,7 @@ SELECT
                     "xx1": 51,
                     "yy1": 5
                 };
-                valActual = await dbExecAndReturnFirstRow({
+                valActual = await dbExecAndReturnLastRow({
                     db,
                     sql: (`
 SELECT
@@ -2646,7 +2737,7 @@ date close
                         priceClose: Number(elem[1])
                     };
                 });
-                valActual = await dbExecAndReturnFirstTable({
+                valActual = await dbExecAndReturnLastTable({
                     bindList: {
                         testDataSpx
                     },